Normalize the line endings
Some files had been commited with CRLF, which causes different trouble under
Linux and Windows.
Normalize all of them.

Change .gitattributes to allow LF in the repo and CRLF in the worktree for
Winndows.

Since the attribute parser in Git does not handle TwinCAT/ as we would need it,
add TwinCAT/.gitattribute to prevent all files here from eol conversion.