FIX: TwinCAT motion library compatibility
This line checks if a struct is of a specific type and enters an
appropriate code branch for it. The check previously failed if the struct
was defined in a TwinCAT library instead of in a cloned subfolder because
the type name gets the library name prepended. This changes the check to
see if the type string contains a specific substring instead of checking
that it matches exactly, allowing us to use our TwinCAT code in library
form without disrupting old use of the module.