Module does not retry to connect on Windows
Constructor of adsAsynPortDriver calls adsUnlock before adsLock. This invokes std::mutex::unlock() with no preceding std::mutex::lock(). According to documentation this results to undefined behavior. While this does not seem to have any effect on Linux, it causes a deadlock-like behavior on Windows when cyclicThread attempts to lock the mutex. As a consequence, the module does not retry to connect.