Add support for 64 bit variables (not arrays)
Support the EPICS 64 bit integer types.
Support for arrays needs to be added (and tested) later.

Make it possible to use an "old asyn/base" which does not support
64 bit integers, see
"#ifndef NO_ADS_ASYN_ASYNPARAMINT64"
every here and there.
However, we do not have Makefile knob yet, that can be added later,
if needed.